Actually improve compatibility with old numpy
I mistakenly tested the previous fix with the new numpy, so it was too easy. This one works.
I'm testing only once if the numpy version accepts the bare "axis" argument, and use the new machinery only in this case.
I do this because this is slower than the numpy implementation.
Feel free to reject this change, and go back to the old behavior. My previous PR was rather ineffective :)